+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
-## Unreleased changes (2026-03-15)
+## Unreleased changes (2026-03-16)
 
-* refactor: testing of solver runs [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
+* tests: fix vampire static args [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
+* refactor: testing of solver runs [[details](https://github.com/cbboyan/solverpy/commit/4674631) | [browse](https://github.com/cbboyan/solverpy/tree/4674631)]
 * tests: prover9 and lash inputs [[details](https://github.com/cbboyan/solverpy/commit/de5909b) | [browse](https://github.com/cbboyan/solverpy/tree/de5909b)]
 * fix: Fixed search for SZS status [[details](https://github.com/cbboyan/solverpy/commit/9d5742b) | [browse](https://github.com/cbboyan/solverpy/tree/9d5742b)]
 * fix: ATP evaluation for Enigma models. [[details](https://github.com/cbboyan/solverpy/commit/2b47e01) | [browse](https://github.com/cbboyan/solverpy/tree/2b47e01)]